refactor backends code; refactor auth code
