+++ apps/core/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoreConfig(AppConfig):
+    name = 'core'
+    verbose_name = 'SSO Core'
