+++ apps/core/backends/__init__.py
@@ -1,0 +1,6 @@
+from apps.core.backends.auth import *
+from apps.core.backends.service import *
+from apps.core.backends.signup import *
+from apps.core.backends.token import *
+from apps.core.backends.user import *
+from apps.core.backends.util import *
