+++ apps/core/backends/util.py
@@ -1,0 +1,55 @@
+from django.conf import settings
+from django.contrib.auth.models import User
+from urllib.parse import urlparse
+import requests
+import re
+
+
+social_name_map = {
+    'FB': 'facebook',
+    'TW': 'twitter',
+    'KAIST': 'kaist',
+}
+
+
+# {male, female, etc} -> {M, F, E}
+def parse_gender(gender):
+    if gender == 'male':
+        return '*M'
+    elif gender == 'female':
+        return '*F'
+    return gender
+
+
+# make clean url; / for all external urls
+def get_clean_url(url):
+    hostname = urlparse(url).hostname
+    print(hostname, settings.DOMAIN)
+    if hostname != settings.DOMAIN:
+        return '/'
+    return url
+
+
+# give pretty name for social account
+def get_social_name(type):
+    return social_name_map.get(type, type)
+
+
+# check given email is available or not
+def validate_email(email, exclude=''):
+    if not re.match(r'[^@]+@[^@]+\.[^@]+', email):
+        return False
+
+    return User.objects.filter(email=email) \
+        .exclude(email=exclude).count() == 0
+
+
+# validate reCAPTCHA
+def validate_recaptcha(response):
+    data = {
+        'secret': settings.RECAPTCHA_SECRET,
+        'response': response
+    }
+    resp = requests.post('https://www.google.com/recaptcha/api/siteverify',
+                         data=data).json()
+    return resp['success']
