+++ apps/core/views/auth.py
@@ -1,19 +1,20 @@
 from django.conf import settings
 from django.contrib import auth
-from django.core.exceptions import PermissionDenied
-from django.core.mail import send_mail
 from django.shortcuts import render, redirect
 from django.utils import timezone
-from apps.core.backends import get_username, \
-    init_fb, init_tw, auth_fb, auth_tw, auth_kaist
+from apps.core.backends import (
+    auth_fb_init, auth_fb_callback,
+    auth_tw_init, auth_tw_callback,
+    auth_kaist_init, auth_kaist_callback,
+    get_clean_url, get_social_name,
+)
 from apps.core.models import Notice, Service
 from urllib.parse import urlparse, parse_qs
 import logging
 
 
 logger = logging.getLogger('sso.core.auth')
-account_logger = logging.getLogger('sso.core.account')
-profile_logger = logging.getLogger('sso.core.profile')
+profile_logger = logging.getLogger('sso.profile')
 
 
 # /login/
@@ -29,44 +30,28 @@ def login(request):
         request.session['next'] = request.GET['next']
 
     query_dict = parse_qs(urlparse(request.session.get('next', '/')).query)
-    service = Service.objects.filter(name=query_dict.get('client_id', [''])[0]).first()
-    srv_name = service.alias if service else ''
+    service_name = query_dict.get('client_id', [''])[0]
+    service = Service.objects.filter(name=service_name).first()
+    service_alias = service.alias if service else ''
 
     if request.method == 'POST':
-        email = request.POST.get('email', 'none')
-        password = request.POST.get('password', 'asdf')
-
-        username = get_username(email)
-        user = auth.authenticate(username=username, password=password)
-        if user and not user.is_active:
-            logger.info('login.reject', {'r': request, 'uid': username})
-            raise PermissionDenied()
-        elif user:
+        email = request.POST.get('email', 'null@sso.sparcs.org')
+        password = request.POST.get('password', 'unknown')
+        user = auth.authenticate(request=request,
+                                 email=email,
+                                 password=password)
+        if user:
             request.session.pop('info_signup', None)
             auth.login(request, user)
-            logger.info('login.success', {'r': request})
-
-            if not settings.DEBUG and user.is_staff:
-                title = '[SPARCS SSO] Staff Login'
-                message = 'time:%s; id:%s; ip:%s;'
-                emails = map(lambda x: x[1], settings.ADMINS)
-                time = timezone.now()
-                ip = request.META.get('REMOTE_ADDR', '0.0.0.0')
-                send_mail(title, '', 'noreply@sso.sparcs.org', emails,
-                          html_message=message % (time, username, ip))
-
-            if user.profile.activate():
-                account_logger.warning('activate', {'r': request})
 
             nexturl = request.session.pop('next', '/')
-            return redirect(nexturl)
-        else:
-            logger.info('login.fail', {'r': request, 'uid': username})
-            request.session['result_login'] = 1
+            return redirect(get_clean_url(nexturl))
+
+        request.session['result_login'] = 1
 
     context = {
         'notice': notice,
-        'service': srv_name,
+        'service': service_alias,
         'fail': request.session.pop('result_login', ''),
         'kaist_enabled': settings.KAIST_APP_ENABLED,
     }
@@ -78,38 +63,36 @@ def logout(request):
     if request.method != 'POST' or not request.user.is_authenticated():
         return redirect('/')
 
-    logger.info('logout', {'r': request})
     auth.logout(request)
     return render(request, 'account/logout.html')
 
 
 # /login/{fb,tw,kaist}/, /connect/{fb,tw,kaist}/, /renew/kaist/
 def init(request, mode, type):
-    is_authed = request.user.is_authenticated()
-    if (mode == 'LOGIN' and is_authed) or \
-       (mode == 'CONN' and not is_authed) or \
-       (mode == 'RENEW' and not is_authed):
+    if request.method != 'POST':
         return redirect('/')
 
-    method = request.method
-    if (mode == 'CONN' and method != 'POST') or \
-       (mode == 'RENEW' and method != 'POST'):
-        return redirect('/account/profile/')
+    # disable login for authed user
+    # disable connect / renew for non authed user
+    is_authed = request.user.is_authenticated()
+    if (mode == 'LOGIN' and is_authed) or \
+       (mode in ['CONN', 'RENEW'] and not is_authed):
+        return redirect('/')
 
-    if is_authed and request.user.profile.test_only and \
-            (mode == 'CONN' or mode == 'RENEW'):
+    # disable manual connect / renew for test user
+    if mode in ['CONN', 'RENEW'] and request.user.profile.test_only:
         return redirect('/account/profile/')
 
     request.session['info_auth'] = {'mode': mode, 'type': type}
     callback_url = request.build_absolute_uri('/account/callback/')
 
-    url = ''
     if type == 'FB':
-        url = init_fb(callback_url)
+        url = auth_fb_init(callback_url)
     elif type == 'TW':
-        url, request.session['request_token'] = init_tw(callback_url)
+        url, token = auth_tw_init(callback_url)
+        request.session['request_token'] = token
     elif type == 'KAIST':
-        url = 'https://ksso.kaist.ac.kr/iamps/requestLogin.do'
+        url = auth_kaist_init()
     return redirect(url)
 
 
@@ -120,21 +103,24 @@ def callback(request):
         return redirect('/')
 
     mode, type = auth['mode'], auth['type']
-    user, profile = None, None
     if type == 'FB':
         code = request.GET.get('code')
         callback_url = request.build_absolute_uri('/account/callback/')
-        profile, info = auth_fb(code, callback_url)
+        profile, info = auth_fb_callback(code, callback_url)
     elif type == 'TW':
         tokens = request.session.get('request_token')
         verifier = request.GET.get('oauth_verifier')
-        profile, info = auth_tw(tokens, verifier)
+        profile, info = auth_tw_callback(tokens, verifier)
     elif type == 'KAIST':
         token = request.COOKIES.get('SATHTOKEN')
-        profile, info = auth_kaist(token)
+        profile, info = auth_kaist_callback(token)
 
-    userid = info['userid'] if info else 'none'
-    logger.info('%s: id=%s' % (type.lower(), userid), {'r': request, 'hide': True})
+    userid = info['userid'] if info else 'unknown'
+    type_str = get_social_name(type)
+    logger.info('social.{}: id={}'.format(type_str, userid), {
+        'r': request,
+        'hide': True,
+    })
     user = profile.user if profile else None
 
     if mode == 'LOGIN':
@@ -154,25 +140,26 @@ def callback_login(request, type, user, info):
         request.session['result_login'] = 2
         return redirect('/account/login/')
 
+    # no such user; go to signup page
     if not user:
         request.session['info_signup'] = {'type': type, 'profile': info}
         response = redirect('/account/signup/social/')
-        response.delete_cookie('SATHTOKEN')
         return response
 
+    user = auth.authenticate(request=request, user=user)
+    # critical logic fail
+    if not user:
+        request.session['result_login'] = 2
+        return redirect('/account/login/')
+
+    # normal login
     request.session.pop('info_signup', None)
     if type == 'KAIST':
         user.profile.set_kaist_info(info)
 
-    user.backend = 'django.contrib.auth.backends.ModelBackend'
     auth.login(request, user)
-    logger.info('login.success', {'r': request})
-
-    if user.profile.activate():
-        account_logger.warning('activate', {'r': request})
-
     nexturl = request.session.pop('next', '/')
-    return redirect(nexturl)
+    return redirect(get_clean_url(nexturl))
 
 
 # from /callback/
@@ -195,13 +182,14 @@ def callback_conn(request, type, user, info):
     profile.save()
     request.session['result_con'] = result_con
 
-    userid = info['userid'] if info else 'none'
+    userid = info['userid'] if info else 'unknown'
+    type_str = get_social_name(type)
     if result_con == 0:
-        profile_logger.warning('connect.success: type=%s,id=%s'
-                               % (type.lower(), userid), {'r': request})
+        profile_logger.warning('social.connect.success: type={},id={}'.format(
+            type_str, userid), {'r': request})
     else:
-        profile_logger.warning('connect.fail: type=%s,id=%s'
-                               % (type.lower(), userid), {'r': request})
+        profile_logger.warning('social.connect.fail: type={},id={}'.format(
+            type_str, userid), {'r': request})
 
     return redirect('/account/profile/')
 
@@ -219,11 +207,14 @@ def callback_renew(request, type, user, info):
         user.profile.set_kaist_info(info)
 
     request.session['result_con'] = result_con
+
+    userid = info['userid'] if info else 'unknown'
+    type_str = get_social_name(type)
     if result_con == 0:
-        profile_logger.warning('renew.success: type=%s,id=%s'
-                               % (type.lower(), info['userid']), {'r': request})
+        profile_logger.warning('social.update.success: type={},id={}'.format(
+            type_str, userid), {'r': request})
     else:
-        profile_logger.warning('renew.fail: type=%s,id=%s'
-                               % (type.lower(), info['userid']), {'r': request})
+        profile_logger.warning('social.fail.success: type={},id={}'.format(
+            type_str, userid), {'r': request})
 
     return redirect('/account/profile/')
